chore: split lotteryrisk.go into focused files [stringer-1gs.4]
Split the 860-line lotteryrisk.go into three focused files:
- lotteryrisk.go: core collector struct, Collect method, directory
  discovery, blame/commit walking
- lotteryrisk_ownership.go: ownership math, lottery risk computation,
  signal building, anonymization
- lotteryrisk_reviews.go: GitHub review participation analysis

Pure refactor — no behavior changes, all tests pass unchanged.
