+++ internal/collectors/lotteryrisk.go
@@ -13,7 +13,6 @@ import (
 
 	"github.com/go-git/go-git/v5"
 	"github.com/go-git/go-git/v5/plumbing/object"
-	"github.com/google/go-github/v68/github"
 
 	"github.com/davetashner/stringer/internal/collector"
 	"github.com/davetashner/stringer/internal/gitcli"
@@ -471,167 +470,6 @@ func recencyDecay(daysOld float64) float64 {
 	return math.Exp(-math.Ln2 / float64(decayHalfLifeDays) * daysOld)
 }
 
-// ownershipFraction computes a weighted ownership score from blame-line and
-// commit-weight contributions. Returns a value in [0, 1].
-func ownershipFraction(blameLines, totalBlameLines int, commitWeight, totalCommitWeight float64) float64 {
-	var blameFrac float64
-	if totalBlameLines > 0 {
-		blameFrac = float64(blameLines) / float64(totalBlameLines)
-	}
-	var commitFrac float64
-	if totalCommitWeight > 0 {
-		commitFrac = commitWeight / totalCommitWeight
-	}
-	return blameFrac*blameWeight + commitFrac*commitWeightFraction
-}
-
-// computeLotteryRisk calculates the lottery risk for a directory: the minimum
-// number of authors whose combined ownership exceeds 50%.
-func computeLotteryRisk(own *dirOwnership) int {
-	if len(own.Authors) == 0 {
-		return 0
-	}
-
-	// Compute combined ownership per author.
-	totalBlameLines := own.TotalLines
-	totalCW := totalCommitWeight(own)
-
-	type authorOwnership struct {
-		Name      string
-		Ownership float64
-	}
-
-	var authors []authorOwnership
-	for name, stats := range own.Authors {
-		ownership := ownershipFraction(stats.BlameLines, totalBlameLines, stats.CommitWeight, totalCW)
-		authors = append(authors, authorOwnership{Name: name, Ownership: ownership})
-	}
-
-	// Sort by ownership descending (highest first).
-	sort.Slice(authors, func(i, j int) bool {
-		if authors[i].Ownership != authors[j].Ownership {
-			return authors[i].Ownership > authors[j].Ownership
-		}
-		return authors[i].Name < authors[j].Name // deterministic tie-break
-	})
-
-	// Count how many authors are needed to exceed 50%.
-	cumulative := 0.0
-	for i, a := range authors {
-		cumulative += a.Ownership
-		if cumulative > ownershipMajority {
-			return i + 1
-		}
-	}
-
-	return len(authors)
-}
-
-// totalCommitWeight sums all authors' commit weights in a directory.
-func totalCommitWeight(own *dirOwnership) float64 {
-	var total float64
-	for _, stats := range own.Authors {
-		total += stats.CommitWeight
-	}
-	return total
-}
-
-// buildDirectoryOwnership converts internal dirOwnership into the exported
-// DirectoryOwnership metrics type.
-func buildDirectoryOwnership(own *dirOwnership) DirectoryOwnership {
-	totalBlameLines := own.TotalLines
-	totalCW := totalCommitWeight(own)
-
-	var authors []AuthorShare
-	for name, stats := range own.Authors {
-		ownership := ownershipFraction(stats.BlameLines, totalBlameLines, stats.CommitWeight, totalCW)
-		authors = append(authors, AuthorShare{Name: name, Ownership: ownership})
-	}
-
-	sort.Slice(authors, func(i, j int) bool {
-		if authors[i].Ownership != authors[j].Ownership {
-			return authors[i].Ownership > authors[j].Ownership
-		}
-		return authors[i].Name < authors[j].Name
-	})
-
-	return DirectoryOwnership{
-		Path:        own.Path,
-		LotteryRisk: own.LotteryRisk,
-		Authors:     authors,
-		TotalLines:  own.TotalLines,
-	}
-}
-
-// buildLotteryRiskSignal constructs a RawSignal for a low-lottery-risk directory.
-// If anon is non-nil, author names are anonymized.
-func buildLotteryRiskSignal(own *dirOwnership, anon *nameAnonymizer) signal.RawSignal {
-	// Find primary author (highest ownership).
-	totalBlameLines := own.TotalLines
-	totalCW := totalCommitWeight(own)
-
-	type authorPct struct {
-		Name string
-		Pct  float64
-	}
-
-	var authors []authorPct
-	for name, stats := range own.Authors {
-		pct := ownershipFraction(stats.BlameLines, totalBlameLines, stats.CommitWeight, totalCW) * 100
-		displayName := name
-		if anon != nil {
-			displayName = anon.anonymize(name)
-		}
-		authors = append(authors, authorPct{Name: displayName, Pct: pct})
-	}
-
-	// Sort by percentage descending, then by name for determinism.
-	sort.Slice(authors, func(i, j int) bool {
-		if authors[i].Pct != authors[j].Pct {
-			return authors[i].Pct > authors[j].Pct
-		}
-		return authors[i].Name < authors[j].Name
-	})
-
-	primary := authors[0]
-
-	// Build description with top authors.
-	var descParts []string
-	descParts = append(descParts, fmt.Sprintf("Lottery risk: %d", own.LotteryRisk))
-	descParts = append(descParts, "Top authors:")
-	for _, a := range authors {
-		if a.Pct < 1.0 {
-			break // skip negligible contributors
-		}
-		descParts = append(descParts, fmt.Sprintf("  - %s: %.0f%%", a.Name, a.Pct))
-	}
-
-	confidence := lotteryRiskConfidence(own.LotteryRisk)
-
-	return signal.RawSignal{
-		Source:      "lotteryrisk",
-		Kind:        "low-lottery-risk",
-		FilePath:    own.Path,
-		Line:        0,
-		Title:       fmt.Sprintf("Low lottery risk: %s (lottery risk %d, primary: %s %.0f%%)", own.Path, own.LotteryRisk, primary.Name, primary.Pct),
-		Description: strings.Join(descParts, "\n"),
-		Confidence:  confidence,
-		Tags:        []string{"low-lottery-risk"},
-	}
-}
-
-// lotteryRiskConfidence maps lottery risk to confidence score per DR-006.
-func lotteryRiskConfidence(riskScore int) float64 {
-	switch {
-	case riskScore <= 1:
-		return 0.8
-	case riskScore == 2:
-		return 0.5
-	default:
-		return 0.3
-	}
-}
-
 // findOwningDir returns the most specific directory in the ownership map
 // that contains the given file path, or empty string if none match.
 func findOwningDir(relPath string, ownership map[string]*dirOwnership) string {
@@ -660,200 +498,6 @@ func findOwningDir(relPath string, ownership map[string]*dirOwnership) string {
 // isSourceExtension returns true for file extensions considered source code.
 func isSourceExtension(ext string) bool {
 	return sourceExtensions[ext]
-}
-
-// nameAnonymizer provides stable, deterministic anonymization of author names.
-// The same real name always maps to the same label within a single scan.
-type nameAnonymizer struct {
-	mapping map[string]string
-	next    int
-}
-
-// newNameAnonymizer creates a new anonymizer.
-func newNameAnonymizer() *nameAnonymizer {
-	return &nameAnonymizer{mapping: make(map[string]string)}
-}
-
-// anonymize returns a stable anonymous label for the given name.
-func (a *nameAnonymizer) anonymize(name string) string {
-	if label, ok := a.mapping[name]; ok {
-		return label
-	}
-	label := contributorLabel(a.next)
-	a.mapping[name] = label
-	a.next++
-	return label
-}
-
-// contributorLabel returns "Contributor A", "Contributor B", ..., "Contributor Z",
-// "Contributor AA", "Contributor AB", etc.
-func contributorLabel(id int) string {
-	if id < 26 {
-		return "Contributor " + string(rune('A'+id))
-	}
-	// For id >= 26: AA=26, AB=27, ..., AZ=51, BA=52, ...
-	first := (id / 26) - 1
-	second := id % 26
-	return "Contributor " + string(rune('A'+first)) + string(rune('A'+second))
-}
-
-// resolveAnonymize determines whether author names should be anonymized based
-// on the mode ("always", "never", "auto") and the repository visibility.
-func resolveAnonymize(ctx context.Context, ghCtx *githubContext, mode string) bool {
-	switch mode {
-	case "always":
-		return true
-	case "never":
-		return false
-	case "auto", "":
-		// Auto mode: anonymize if the repo is public.
-		if ghCtx == nil {
-			return false // no API available, default to not anonymizing
-		}
-		repo, _, err := ghCtx.API.GetRepository(ctx, ghCtx.Owner, ghCtx.Repo)
-		if err != nil || repo == nil {
-			return false // can't determine visibility, default to not anonymizing
-		}
-		// Public repos → anonymize; private repos → don't.
-		return !repo.GetPrivate()
-	default:
-		return false
-	}
-}
-
-// reviewParticipation tracks review activity in a directory.
-type reviewParticipation struct {
-	Reviewers map[string]int // reviewer login → review count
-	Authors   map[string]int // PR author login → PR count
-}
-
-// fetchReviewParticipation fetches merged PRs and their reviews from GitHub,
-// then maps review activity to directories based on changed files.
-func fetchReviewParticipation(ctx context.Context, ghCtx *githubContext, ownership map[string]*dirOwnership, maxPRs int) (map[string]*reviewParticipation, error) {
-	result := make(map[string]*reviewParticipation)
-
-	// Fetch recently merged PRs.
-	opts := &github.PullRequestListOptions{
-		State:     "closed",
-		Sort:      "updated",
-		Direction: "desc",
-		ListOptions: github.ListOptions{
-			PerPage: 100,
-		},
-	}
-
-	fetched := 0
-	for {
-		if err := ctx.Err(); err != nil {
-			return nil, err
-		}
-
-		prs, resp, err := ghCtx.API.ListPullRequests(ctx, ghCtx.Owner, ghCtx.Repo, opts)
-		if err != nil {
-			return nil, fmt.Errorf("listing merged PRs for review analysis: %w", err)
-		}
-
-		for _, pr := range prs {
-			if !pr.GetMerged() {
-				continue
-			}
-			if fetched >= maxPRs {
-				return result, nil
-			}
-			fetched++
-
-			if err := ctx.Err(); err != nil {
-				return nil, err
-			}
-
-			// Fetch reviews for this PR.
-			reviews, reviewErr := fetchAllReviews(ctx, ghCtx.API, ghCtx.Owner, ghCtx.Repo, pr.GetNumber())
-			if reviewErr != nil {
-				continue // skip PRs with review fetch errors
-			}
-
-			// Fetch files changed in this PR.
-			files, _, filesErr := ghCtx.API.ListPullRequestFiles(ctx, ghCtx.Owner, ghCtx.Repo, pr.GetNumber(), &github.ListOptions{PerPage: 100})
-			if filesErr != nil {
-				continue // skip PRs with file fetch errors
-			}
-
-			// Determine which directories this PR touches.
-			touchedDirs := make(map[string]bool)
-			for _, f := range files {
-				dir := findOwningDir(f.GetFilename(), ownership)
-				if dir != "" {
-					touchedDirs[dir] = true
-				}
-			}
-
-			// Attribute reviews and authorship to directories.
-			for dir := range touchedDirs {
-				if result[dir] == nil {
-					result[dir] = &reviewParticipation{
-						Reviewers: make(map[string]int),
-						Authors:   make(map[string]int),
-					}
-				}
-
-				result[dir].Authors[pr.GetUser().GetLogin()]++
-
-				for _, review := range reviews {
-					state := strings.ToUpper(review.GetState())
-					if state == "APPROVED" || state == "CHANGES_REQUESTED" {
-						result[dir].Reviewers[review.GetUser().GetLogin()]++
-					}
-				}
-			}
-		}
-
-		if resp.NextPage == 0 {
-			break
-		}
-		opts.Page = resp.NextPage
-	}
-
-	return result, nil
-}
-
-// buildReviewConcentrationSignals produces signals for directories where a
-// single reviewer handles more than 70% of all reviews.
-// If anon is non-nil, reviewer names are anonymized.
-func buildReviewConcentrationSignals(reviewData map[string]*reviewParticipation, anon *nameAnonymizer) []signal.RawSignal {
-	var signals []signal.RawSignal
-
-	for dir, rp := range reviewData {
-		totalReviews := 0
-		for _, count := range rp.Reviewers {
-			totalReviews += count
-		}
-		if totalReviews < 3 {
-			continue // not enough data to draw conclusions
-		}
-
-		for reviewer, count := range rp.Reviewers {
-			fraction := float64(count) / float64(totalReviews)
-			if fraction > reviewConcentrationThreshold {
-				displayName := reviewer
-				if anon != nil {
-					displayName = anon.anonymize(reviewer)
-				}
-
-				signals = append(signals, signal.RawSignal{
-					Source:      "lotteryrisk",
-					Kind:        "review-concentration",
-					FilePath:    dir,
-					Line:        0,
-					Title:       fmt.Sprintf("Review bottleneck: %s reviews %.0f%% of PRs in %s", displayName, fraction*100, dir),
-					Description: fmt.Sprintf("Reviewer %s handled %d of %d reviews (%.0f%%) in %s. Consider distributing review responsibility to reduce knowledge silos.", displayName, count, totalReviews, fraction*100, dir),
-					Confidence:  0.6,
-					Tags:        []string{"review-concentration"},
-				})
-			}
-		}
-	}
-
-	return signals
 }
 
 // Metrics returns structured ownership data for all analyzed directories.
